// frontend/projects/smudgy/src/app/models/network/events.ts
export enum Events {
  Register = 'register',
  CreateSession = 'create-session',
  JoinSession = 'join-session',
  PlayerLeaveSession = 'player-leave-session',
  PlayerJoinSession = 'player-join-session',
  UpdateSessionConfiguration = 'update-session-configuration',
  PlayerList = 'player-list',
  StartGame = 'start-game',
}